requirements.txt
pandas
numpy
requests
ccxt
pycoingecko
pyyaml
scikit-learn
python-binance
yfinance
Flask
torch
matplotlib
hvac>=2.3.0
pysui

